# API/Specification/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample Blog server where the comments manipulation occurs.",
    "version": "1.0.0",
    "title": "Comments Store",
    "contact": {
      "email": "prasath.balu13@gmail.com"
    }
  },
  "host": "localhost:3000",
  "tags": [
    {
      "name": "comment",
      "description": "Blog Comments"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/comments": {
      "post": {
        "tags": [
          "comment"
        ],
        "summary": "Add a new comment for the blog",
        "description": "",
        "operationId": "addComment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Please validate the input content"
          }
        }
      },
      "get": {
        "tags": [
          "comment"
        ],
        "summary": "Get all comments",
        "description": "Comments will be provided based on the parent Id",
        "operationId": "GetComments",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          },
          "400": {
            "description": "Invalid parentId"
          }
        }
      }
    },
    "/comments/parent": {
      "get": {
        "tags": [
          "comment"
        ],
        "summary": "Find all the comments by parent id",
        "description": "Commnents will be provided based on the parent Id",
        "operationId": "findcommentById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "description": "value that need to be considered for filter",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          },
          "400": {
            "description": "Invalid parent value"
          }
        }
      }
    }
  },
  "definitions": {
    "Comment": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "format": "objectId"
        },
        "comment": {
          "type": "string"
        },
        "parentId": {
          "type": "string",
          "format": "objectId"
        },
        "blogId": {
          "type": "string",
          "format": "objectId"
        },
        "whenCreated": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "description": "For Internal server error",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}